objects/draw: merge Object_DrawSpriteItem in TRX
