+++ docs/tr2/CHANGELOG.md
@@ -52,6 +52,7 @@
 - changed the Pause key to no longer work when Lara's dead (similar to TR1)
 - changed the `-q`/`--quiet` argument to no longer silence warnings
 - changed the `Remember Guns between Levels` option to also apply to whether or not Lara starts with those guns equipped
+- changed sprites to respect the water tint if placed underwater
 - removed the following game flow options:
     - `cmd_init`
     - `cmd_title`
