+++ src/tr1/game/objects/common.c
@@ -6,16 +6,6 @@
 #include <libtrx/game/matrix.h>
 #include <libtrx/game/output.h>
 
-void Object_DrawSpriteItem(const ITEM *const item)
-{
-    const RGB_F tint = Output_GetTint();
-    Output_DrawSprite(
-        item->interp.result.pos.x, item->interp.result.pos.y,
-        item->interp.result.pos.z,
-        Object_Get(item->object_id)->mesh_idx - item->frame_num,
-        item->shade.value_1 < 0 ? SHADE_NEUTRAL : item->shade.value_1, tint);
-}
-
 void Object_DrawPickupItem(const ITEM *const item)
 {
     if (!g_Config.visuals.enable_3d_pickups) {
